requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.5
django-filter==21.1
django-seed==0.3.1
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==9.9.0
filelock==3.4.0
flake8==4.0.1
identify==2.4.0
idna==3.3
inflection==0.5.1
ipython==7.30.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
nodeenv==1.6.0
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pre-commit==2.16.0
prompt-toolkit==3.0.23
psycopg2==2.9.2
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
toposort==1.7
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.2.5
